Update index.html
The current project is organized so that the index.html for One Direction is in it's own folder, thereby making all the paths to your .css and .js files invalid (it looks for them inside the folder that 1D's index.html is in instead of the root folder). I added .. in front of all of the paths to go up a directory, which then makes your file paths correctly drill down to the expected file.